ingestion/requirements.txt
functions-framework==3.10.1
google-cloud-storage==3.10.0
pdfplumber==0.11.9
google-cloud-firestore